refactor: restructure meta->lib directory
