+++ package.json
@@ -15,6 +15,7 @@
 		"formik-material-ui": "^2.0.0-beta.1",
 		"history": "^4.10.1",
 		"react": "^16.13.1",
+		"react-color": "^2.18.1",
 		"react-dom": "^16.13.1",
 		"react-emoji-render": "^1.2.4",
 		"react-flip-move": "^3.0.4",
@@ -53,6 +54,7 @@
 		"@types/lodash": "^4.14.152",
 		"@types/node": "^14.0.3",
 		"@types/react": "^16.9.35",
+		"@types/react-color": "^3.0.1",
 		"@types/react-dom": "^16.9.8",
 		"@types/react-router": "^5.1.7",
 		"@types/react-router-dom": "^5.1.5",
