+++ src/components/Formik/AutoSaveForm.tsx
@@ -1,6 +1,6 @@
 import { Form, Formik, FormikHelpers, useFormikContext } from 'formik';
+import { Time } from 'lib/util/skyraUtils';
 import { debounce } from 'lodash';
-import { Time } from 'meta/constants';
 import React, { Fragment, PropsWithChildren, useCallback, useEffect } from 'react';
 import { ObjectSchema } from 'yup';
