+++ src/components/GeneralPage.tsx
@@ -18,8 +18,8 @@ import SkyraLogo from 'assets/skyraLogo';
 import Footer from 'components/Footer';
 import MobileNavMenu from 'components/MobileNavMenu';
 import UserMenu from 'components/UserMenu';
-import { oauthURL } from 'meta/constants';
-import { navigate, syncUser } from 'meta/util';
+import { oauthURL } from 'lib/util/constants';
+import { navigate, syncUser } from 'lib/util/util';
 import React, { PropsWithChildren, useEffect } from 'react';
 import { Else, If, Then, When } from 'react-if';
 import { useGlobal } from 'reactn';
