+++ src/components/Root.tsx
@@ -3,8 +3,8 @@ import { StylesProvider, ThemeProvider } from '@material-ui/styles';
 import AuthenticatedRoute from 'components/AuthenticatedRoute';
 import ErrorBoundary from 'components/ErrorBoundary';
 import RedirectRoute from 'components/RedirectRoute';
-import { history, oauthURL, serverURL } from 'meta/constants';
-import theme from 'meta/theme';
+import { history, oauthURL, serverURL } from 'lib/util/constants';
+import theme from 'lib/theme';
 import CommandsPage from 'pages/Commands';
 import DashboardRootPage from 'pages/Dashboard/Root';
 import HomePage from 'pages/HomePage';
