+++ src/components/UserMenu.tsx
@@ -13,7 +13,8 @@ import Typography from '@material-ui/core/Typography';
 import LogoutIcon from '@material-ui/icons/Eject';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import SyncIcon from '@material-ui/icons/Sync';
-import { displayAvatarURL, logOut, syncUser } from 'meta/util';
+import { displayAvatarURL } from 'lib/util/skyraUtils';
+import { logOut, syncUser } from 'lib/util/util';
 import React, { useEffect, useRef, useState } from 'react';
 import { useGlobal } from 'reactn';
 import Tooltip from './Tooltip';
