+++ src/lib/types/GuildSettings.ts
