+++ src/pages/Dashboard/Root.tsx
@@ -42,9 +42,10 @@ import GuildIcon from 'components/GuildIcon';
 import Tooltip from 'components/Tooltip';
 import UserMenu from 'components/UserMenu';
 import deepMerge, { Options as DeepMergeOptions } from 'deepmerge';
-import { FlattenedGuild } from 'meta/typings/ApiData';
-import { GuildSettings, SettingsPageProps } from 'meta/typings/GuildSettings';
-import { authedFetch, navigate, noOp } from 'meta/util';
+import { FlattenedGuild } from 'lib/types/ApiData';
+import { GuildSettings, SettingsPageProps } from 'lib/types/GuildSettings';
+import { noOp } from 'lib/util/skyraUtils';
+import { authedFetch, navigate } from 'lib/util/util';
 import CustomCommandsPage from 'pages/Dashboard/CustomCommands';
 import EventsPage from 'pages/Dashboard/EventsPage';
 import FilterCapitalsPage from 'pages/Dashboard/Filter/Capitals';
